refactor: split ImagePicker into grid, external panel, buffer, utils
675 lines held the modal shell, the Payload results grid, two provider
search panels with an inline crop editor, the multi-select buffer, and
upload-identity helpers.

Split into:
- ImagePickerGrid.tsx                 results grid (was already a separate
                                      component, just below the fold)
- ExternalImagePanel.tsx              provider search + crop editor +
                                      shared MultiSelectFooter
- useImagePickerSelectionBuffer.ts    the ordered-ids + resolution-map buffer
- imagePicker.utils.ts                ref sanitizing / upload identity
- ImagePicker.tsx                     modal shell (675 -> 335 lines)

The two render-prop closures (renderExternalPanel / renderExternalCropEditor)
became real components with explicit props. ExternalImagePanel is generic
over the photo type so Unsplash and Pexels controllers pass through without
casts.

The multi-select footer JSX was duplicated verbatim under the Payload grid
and the provider results; both now render one MultiSelectFooter.

Verified: picker barrel export surface identical (14/14 via TS checker vs
main); tsc error signature identical to baseline; check-boundaries output
identical to main; vitest 333 pass / 3 pre-existing failures, with all 20
picker-related tests passing (useImagePickerData, mediaSet.utils,
FeaturedImagePicker).
